updated flow, still strugling with webclient bug
The intenral web client used by quarkus and vertx has ssl disabled so
all outgoing rest requests are http and this fails the POST requests for
submiting a taskgroup.
This can be solved by programatically initiaing the webclient options
and creating the a java:call and a service inteface that the function would call,
instead of using the rest:post function.

Signed-off-by: Roy Golan <rgolan@redhat.com>